backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
python-dotenv==1.0.0
python-multipart==0.0.6
openai-whisper==20231117
google-generativeai==0.3.2
pydantic==2.5.0
pydantic-settings==2.1.0

